Bad description formating
Sometimes a sjp.pl site is in bad formating.
Are different from standard site.
In this case in meaningOf will be "BAD FORMATING" state.
If someone would like to get any info about meaning this word
should go to site: "sjp.pl/{word}".

name, canBeUsed oraz variant is always good formating,
so it is not neccesary to do someting with this fields.
User should expect "BAD FORMAT" state and do something
to handling that

Closes-issue: Word without meaning https://github.com/GHRik/SjpAPI/issues/6